backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
chromadb==0.5.0
sentence-transformers==3.0.0
anthropic==0.34.0
langchain-text-splitters==0.2.0
pydantic==2.9.0
python-dotenv==1.0.0
